Rename jsonCompiler to libsolc
