+++ CMakeLists.txt
@@ -43,9 +43,10 @@ configure_project(TESTS)
 add_subdirectory(libdevcore)
 add_subdirectory(libevmasm)
 add_subdirectory(libsolidity)
-add_subdirectory(solc)
+add_subdirectory(libsolc)
 
 if (NOT EMSCRIPTEN)
+	add_subdirectory(solc)
 	add_subdirectory(liblll)
 	add_subdirectory(lllc)
 endif()
