+++ libsolc/CMakeLists.txt
@@ -1,0 +1,9 @@
+if (EMSCRIPTEN)
+	set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -s EXPORTED_FUNCTIONS='[\"_compileJSON\",\"_license\",\"_version\",\"_compileJSONMulti\",\"_compileJSONCallback\",\"_compileStandard\"]' -s RESERVED_FUNCTION_POINTERS=20")
+	add_executable(soljson libsolc.cpp)
+	target_link_libraries(soljson PRIVATE solidity)
+else()
+	add_library(libsolc libsolc.cpp)
+	set_target_properties(libsolc PROPERTIES OUTPUT_NAME solc)
+	target_link_libraries(libsolc PRIVATE solidity)
+endif()
